# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Service designed for automatic news parsing",
        "title": "News-Parser",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/api/v1/article/archive/text/{uuid}": {
            "get": {
                "description": "get one archive article",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "article_default"
                ],
                "summary": "get archive article",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "uuid  archive article must be 36 characters",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/app_news-parser_internal_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/app_news-parser_internal_model.ArticleArchive"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "validation errors:\" | `uuid` - article_uuid must be 36 characters long",
                        "schema": {
                            "$ref": "#/definitions/app_news-parser_internal_response.Response"
                        }
                    },
                    "404": {
                        "description": "not found:\" | `not_found_articles` - article not found",
                        "schema": {
                            "$ref": "#/definitions/app_news-parser_internal_response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/article/archive/{category}": {
            "get": {
                "description": "get archive articles by category",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "article_default"
                ],
                "summary": "get archive articles",
                "parameters": [
                    {
                        "type": "string",
                        "description": "food, politics, sport, cloth, electronics, the following categories are valid, as well as all to get all entries",
                        "name": "category",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "offset default 0",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "limit default 50",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "date must be YYYY-MM-DD",
                        "name": "date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/app_news-parser_internal_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/internal_article_default.ResponseCategoryArchive"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "validation errors:\" | `offset` - the offset must be a positive integer | `limit` - the limit must be a positive integer | `category` - this category of articles does not exist | `date` - the date format must be YYYY-MM-DD",
                        "schema": {
                            "$ref": "#/definitions/app_news-parser_internal_response.Response"
                        }
                    },
                    "404": {
                        "description": "not found:\" | `not_found_articles` - article not found",
                        "schema": {
                            "$ref": "#/definitions/app_news-parser_internal_response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/article/today/text/{id}": {
            "get": {
                "description": "get one today article with text",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "article_default"
                ],
                "summary": "get article with text",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id of today's article must be 7 digits",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/app_news-parser_internal_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/app_news-parser_internal_model.ArticleToday"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "validation errors:\" | `id` - id must be 7 digits long",
                        "schema": {
                            "$ref": "#/definitions/app_news-parser_internal_response.Response"
                        }
                    },
                    "404": {
                        "description": "not found:\" | `not_found_articles` - article not found",
                        "schema": {
                            "$ref": "#/definitions/app_news-parser_internal_response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/article/today/{category}": {
            "get": {
                "description": "get today's articles by category",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "article_default"
                ],
                "summary": "get today articles",
                "parameters": [
                    {
                        "type": "string",
                        "description": "food, politics, sport, cloth, electronics, the following categories are valid, as well as all to get all entries",
                        "name": "category",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "offset default 0",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "limit default 50",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "is_articles default true",
                        "name": "is_articles",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "with_text default false",
                        "name": "with_text",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/app_news-parser_internal_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/internal_article_default.ResponseCategoryToday"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "validation errors:\" | `is_articles` - the isArticles must be a boolean is true or false | `with_text` - the withText must be a boolean is true or false | `offset` - the offset must be a positive integer | `limit` - the limit must be a positive integer | `category` - this category of articles does not exist",
                        "schema": {
                            "$ref": "#/definitions/app_news-parser_internal_response.Response"
                        }
                    },
                    "404": {
                        "description": "not found:\" | `not_found_articles` - article not found",
                        "schema": {
                            "$ref": "#/definitions/app_news-parser_internal_response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "app_news-parser_internal_model.ArticleArchive": {
            "type": "object",
            "properties": {
                "articleArchiveUUID": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "header": {
                    "type": "string"
                },
                "isArticle": {
                    "type": "boolean"
                },
                "text": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "app_news-parser_internal_model.ArticleToday": {
            "type": "object",
            "properties": {
                "articleID": {
                    "type": "integer"
                },
                "category": {
                    "type": "string"
                },
                "header": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "app_news-parser_internal_response.Error": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "app_news-parser_internal_response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/app_news-parser_internal_response.Error"
                    }
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "internal_article_default.ResponseCategoryArchive": {
            "type": "object",
            "properties": {
                "header": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "uuidarticle": {
                    "type": "string"
                }
            }
        },
        "internal_article_default.ResponseCategoryToday": {
            "type": "object",
            "properties": {
                "header": {
                    "type": "string"
                },
                "idarticle": {
                    "type": "integer"
                },
                "isArticle": {
                    "type": "boolean"
                },
                "text": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
    }
}